requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
future==0.16.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.0.0
PyJWT==1.6.1
pytz==2018.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
future==0.16.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.0.0
PyJWT==1.6.1
pytz==2018.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
